updating the dev-qa codes, and deleting unused code.
