+++ StellarApp/Protocol/PlanetsMoved.swift
@@ -103,12 +103,6 @@ struct MovedPlanet {
 }
 
 extension SCNNode {
-    static func deepScaleNodes(node: SCNNode, scaleAction: SCNAction) {
-        node.runAction(scaleAction)
-        for child in node.childNodes {
-            deepScaleNode(node: child, scaleAction: scaleAction)
-        }
-    }
     
     static func deepScaleNodes(node: SCNNode, scale: CGFloat) {
         let scaleX = scale * CGFloat(node.scale.x)
@@ -171,20 +165,4 @@ extension SCNNode {
     
 }
 
-struct Sound {
-static var player: AVAudioPlayer?
-   static func playSound(sound : String, format: String) {
-        guard let url = Bundle.main.url(forResource: sound, withExtension: format) else { return }
-        do {
-            try AVAudioSession.sharedInstance().setCategory(AVAudioSession.Category.playback)
-            try AVAudioSession.sharedInstance().setActive(true)
-            
-            player = try AVAudioPlayer(contentsOf: url, fileTypeHint: AVFileType.mp3.rawValue)
-            
-            guard let player = player else { return }
-            player.play()
-        } catch let error {
-            print(error.localizedDescription)
-        }
-    }
-}
+
