+++ StellarApp/Protocol/SoundsFiles.swift
@@ -1,0 +1,28 @@
+//
+//  SoundsFiles.swift
+//  StellarApp
+//
+//  Created by TingxinLi on 5/1/19.
+//  Copyright © 2019 TingxinLi. All rights reserved.
+//
+
+import UIKit
+import ARKit
+
+struct Sound {
+    static var player: AVAudioPlayer?
+    static func playSound(sound : String, format: String) {
+        guard let url = Bundle.main.url(forResource: sound, withExtension: format) else { return }
+        do {
+            try AVAudioSession.sharedInstance().setCategory(AVAudioSession.Category.playback)
+            try AVAudioSession.sharedInstance().setActive(true)
+            
+            player = try AVAudioPlayer(contentsOf: url, fileTypeHint: AVFileType.mp3.rawValue)
+            
+            guard let player = player else { return }
+            player.play()
+        } catch let error {
+            print(error.localizedDescription)
+        }
+    }
+}
